requirements.txt
requests
python-dotenv
upstox-python-sdk
groq